+++ tools/moon/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+#### 🚀 Updates
+
+- Added build from source support.
+
 ## 0.0.2
 
 #### 🎉 Release